# Conflicts:
#	DoAnTotNghiep_MiniatureCrafts/build.gradle
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Controller/Auth/AuthController.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Controller/Product/HomeController.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Controller/user/UsersController.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/DTO/BrandDTO.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/DTO/POSOrderDTO.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Entity/Account.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Entity/Brand.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Entity/ERole.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Entity/POSOrder.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Entity/Variation.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Repository/User/EmployeeRepository.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Repository/User/UserRepository.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Service/Order/OrderService.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/Service/Security/UserService.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/security/jwt/AuthTokenFilter.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/security/services/UserDetailsImpl.java
#	DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/security/services/UserDetailsServiceImpl.java

+++ DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/DTO/BrandDTO.java
@@ -3,12 +3,13 @@ package com.example.DoAnTotNghiep_MiniatureCrafts.DTO;
 import com.example.DoAnTotNghiep_MiniatureCrafts.Entity.Brand;
 
 public class BrandDTO {
-    private Integer ID;
+    private Long ID;
     private String Name;
     private String Note;
     private Boolean Status;
 
-    public BrandDTO(Integer ID, String name, String note, Boolean status) {
+
+    public BrandDTO(Long ID, String name, String note, Boolean status) {
         this.ID = ID;
         Name = name;
         Note = note;
@@ -38,11 +39,11 @@ public class BrandDTO {
 
 
     // Getters and Setters
-    public Integer getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(Integer ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
