+++ DoAnTotNghiep_MiniatureCrafts/src/main/java/com/example/DoAnTotNghiep_MiniatureCrafts/DTO/CategoryDTO.java
@@ -3,7 +3,7 @@ package com.example.DoAnTotNghiep_MiniatureCrafts.DTO;
 import com.example.DoAnTotNghiep_MiniatureCrafts.Entity.Category;
 
 public class CategoryDTO {
-    private Integer ID;
+    private Long ID;
     private String Name;
     private Boolean Status;
 
@@ -14,11 +14,11 @@ public class CategoryDTO {
     }
 
     // Getters and Setters
-    public Integer getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(Integer ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
